backend/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.26.3
azure-storage-blob==12.15.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
detect-delimiter==0.1.1
dnspython==2.3.0
editdistance==0.6.2
efficientnet==1.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
essential-generators==1.0
fastapi==0.94.1
flatbuffers==23.3.3
fonttools==4.39.2
frozenlist==1.3.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
idna==3.4
imageio==2.26.1
imgaug==0.4.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
isodate==0.6.1
Jinja2==3.1.2
Keras-Applications==1.0.8
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.1
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
msrest==0.7.1
multidict==6.0.4
murmurhash==1.0.9
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
openai==1.16.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
# packaging==23.0
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.21
pydantic==1.10.6
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-pptx==0.6.21
PyWavelets==1.4.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
# scipy==1.10.0
shapely==2.0.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
starlette==0.26.1
termcolor==2.2.0
thinc==8.1.9
tifffile==2023.3.15
tqdm==4.65.0
typer==0.7.0
# typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.20.0
validators==0.20.0
wasabi==1.1.1
Werkzeug==2.2.3
wrapt==1.15.0
XlsxWriter==3.0.9
yarl==1.8.2
zipp==3.15.0
# openai==0.27.2
PyPDF4==1.27.0
Aspose.slides==23.5.0
pdf2image==1.16.3
scikit-learn==1.2.2
nltk==3.8.1
pandas==2.0.0
transformers==4.27.4
whoosh==2.7.4
gensim==4.3.1
fuzzywuzzy==0.18.0
elasticsearch==8.7.0
asposeslidescloud==23.4.0
backoff==2.2.1
apscheduler==3.10.1
tzlocal==5.0.1
retry==0.9.2
pyPDF2==3.0.1
langid==1.1.6
azure-search-documents==11.4.0b6
pypdf==3.14.0
# tiktoken==0.4.0
fastapi_mail==1.1.5
PyJWT==2.6.0
bcrypt==4.0.1
fastapi-jwt==0.2.0
passlib==1.7.4
azure-identity==1.12.0
azure-cognitiveservices-speech==1.24.0
azure-keyvault-administration==4.3.0
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.8.0b2
azure-keyvault-secrets==4.7.0
azure-cli==2.53.0
azure-cli-core==2.53.0
azure-cli-telemetry==1.1.0
aspose-words==23.9.0
azure-ai-formrecognizer==3.3.2
aspose-cells-python==23.10.0
databricks-sdk==0.23.0
pymupdf==1.24.0
python-multipart==0.0.9
databricks-connect==13.3.3
SQLAlchemy==2.0.26
databricks-sql-connector==3.0.0
langchain==0.1.8
langchain-community==0.0.24
langchain-openai==0.0.8
flashrank==0.2.6
plotly==5.22.0
plotly-express==0.4.1
kaleido==0.1.*